requirements.txt
certifi
click
Flask
GitPython
idna
itsdangerous
Jinja2
MarkupSafe
pyTelegramBotAPI
requests
Send2Trash
smmap
urllib3
Werkzeug
SQLAlchemy
python-dotenv